% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{breakties_par}
\alias{breakties_par}
\title{make logical parameter which determines if ties should be broken randomly
with parameter stability tests of continuous splitting variables}
\usage{
breakties_par(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{a vector of possible values; only TRUE/FALSE possible (the default)}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make logical parameter which determines if ties should be broken randomly
with parameter stability tests of continuous splitting variables
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.001, alpha_max = 0.1),
   bonferroni_par(),
   size = 10
   )
}
