% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{catsplit_par}
\alias{catsplit_par}
\title{make parameter for method of splitting factors (unordered categorical variables)}
\usage{
catsplit_par(values = c("binary", "multiway"))
}
\arguments{
\item{values}{a vector of catsplit_par values (only 2 options currently available)}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make parameter for method of splitting factors (unordered categorical variables)
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.001, alpha_max = 0.1),
   catsplit_par(),
   size = 10
   )
}
