% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{class_acc}
\alias{class_acc}
\title{classification accuracy}
\usage{
class_acc(observed_y, predicted_y)
}
\arguments{
\item{observed_y}{A numeric vector of observed ys.}

\item{predicted_y}{A numeric vector of predictions of y.}
}
\value{
A double vector of length 1
}
\description{
classification accuracy
}
\examples{
set.seed(1)
set.seed(1)
obs_y <- rbinom(n = 4, size = 1, prob = 0.5)
pred_y <- rbinom(n = 4, size = 1, prob = 0.5)
class_acc(observed_y = obs_y, predicted_y = pred_y)
}
