% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{cross_validate_it}
\alias{cross_validate_it}
\title{fit \code{glmertree()} to a cross-validated data set}
\usage{
cross_validate_it(cv_obj, seed = 713, mod_formula, ...)
}
\arguments{
\item{cv_obj}{vfold_cv—a v-fold cross-validated dataset from \code{rsample::vfold_cv()}}

\item{seed}{integer—starting seed}

\item{mod_formula}{Formula—made from \code{as.Formula()}. uppercase required \verb{:)}}

\item{...}{additional arguments to be passed to \code{glmertree()}}

\item{make_my_tuning_grid}{logical—should default tuning grid be used? (temporary)}
}
\value{
tibble—fit statistics (rmse, mae) for object
}
\description{
fit \code{glmertree()} to a cross-validated data set
}
\examples{
dat <- sim_multilevel()
example_split <- rsample::initial_split(dat)
example_train <- rsample::training(example_split)
example_test  <-  rsample::testing(example_split)
cv <- rsample::vfold_cv(data = example_train, v = 10)

ex_formula <-
   Formula::as.Formula(
      'outcome ~ small_1 |
      (1 | id_vector) |
      small_c_1 + small_c_2 + nuisance_1a + nuisance_c_1a'
      )
fitted <-
   cross_validate_it(
      cv_obj = cv,
      seed = 713,
      mod_formula = ex_formula,
      make_my_tuning_grid = TRUE
      )
}
