% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garden_coords.R
\docType{data}
\name{garden_coords}
\alias{garden_coords}
\title{Data from Lisa Lendway's backyard garden}
\format{
A tibble with 76 rows and 3 variables:
\describe{
  \item{plot}{label of plot where vegetables were planted}
  \item{x}{x coordinate for a vertex of a plot}
  \item{y}{y coordinate for a vertex of a plot}
}
}
\source{
Manually collected by Lisa Lendway, initially in a Google sheet.
}
\usage{
garden_coords
}
\description{
During the summer of 2020, Dr. Lisa Lendway decided it would be fun to collect data on the vegetables she harvested from her backyard garden, which she fondly calls the "jungle garden."  This dataset gives coordinates for the vertices of the plots in the garden.
}
\keyword{datasets}
