% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gardenR-package.R
\docType{package}
\name{gardenR-package}
\alias{gardenR}
\alias{gardenR-package}
\title{gardenR: Data Package With Data From Lisa Lendway's Garden}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This is a data package with datasets recorded manually by Lisa Lendway.
}
\keyword{internal}
