% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_all_counts}
\alias{get_all_counts}
\title{Returns count for all categories}
\usage{
get_all_counts(folder)
}
\description{
Does not include the category 'None', as it is not unique across features.
The categories are sorted according to features F1-F6 and then alphabetically.
Does not include 'complete' as it is at annotator-cell level, not cell level
Does not include alternative type names as their number varies from cell to cell
}
