% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{mae}
\alias{mae}
\title{mean absolute error}
\usage{
mae(observed_y, predicted_y)
}
\arguments{
\item{observed_y}{A numeric vector of observed ys.}

\item{predicted_y}{A numeric vector of predictions of y.}
}
\value{
A double vector of length 1
}
\description{
mean absolute error
}
\examples{
set.seed(1)
obs_y <- rnorm(4)
pred_y <- rnorm(4)
mae(observed_y = obs_y, predicted_y = pred_y)
}
