% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{maxdepth_par}
\alias{maxdepth_par}
\title{make parameter for tuning maximum depth of tree}
\usage{
maxdepth_par(maxdepth_min = 2, maxdepth_max = 12)
}
\arguments{
\item{maxdepth_min}{An integer specifying the minimum value in the parameter grid}

\item{maxdepth_max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for maxdepth
}
\description{
make parameter for tuning maximum depth of tree
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.001, alpha_max = 0.1),
   trim_par(trim_min = 0.1, trim_max = 0.3),
   size = 10
   )
}
