% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{minsize_par}
\alias{minsize_par}
\title{make parameter for tuning minimum (terminal) node size}
\usage{
minsize_par(minsize_min, minsize_max)
}
\arguments{
\item{maxdepth_min}{An integer specifying the minimum value in the parameter grid}

\item{maxdepth_max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for minsize
}
\description{
make parameter for tuning minimum (terminal) node size
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.10, alpha_max = 0.001),
   minsize_par(minsize_min = 200, minsize_max = 500),
   size = 10
   )
}
