% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{mtry_par}
\alias{mtry_par}
\title{make parameter for mtry}
\usage{
mtry_par(mtry_min, mtry_max)
}
\arguments{
\item{mtry_min}{An integer specifying the minimum value in the parameter grid}

\item{mtry_max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for mtry
}
\description{
make parameter for mtry
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   mtry_par(mtry_min = 1, mtry_max = 20),
   minsize_par(minsize_min = 200, minsize_max = 500),
   size = 10
   )
}
