% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{nrep_par}
\alias{nrep_par}
\title{make parameter for \code{nrep}, number of replications in simulation of p-values with L2 statistic}
\usage{
nrep_par(nrep_min, nrep_max)
}
\arguments{
\item{nrep_min}{minimum of nrep range}

\item{nrep_max}{maximum of nrep range}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make parameter for \code{nrep}, number of replications in simulation of p-values with L2 statistic
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.10, alpha_max = 0.001),
   nrep_par(1000, 10000),
   size = 10
   )
}
