% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{ordinal_par}
\alias{ordinal_par}
\title{make parameter for determining how to handle ordered categorical splits}
\usage{
ordinal_par(values = c("chisq", "max", "L2"))
}
\arguments{
\item{values}{a vector of possible values c("chisq", "max", "L2"), see ?mob_control()}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make parameter for determining how to handle ordered categorical splits
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.001, alpha_max = 0.1),
   bonferroni_par(),
   size = 10
   )
}
