% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{prune_par}
\alias{prune_par}
\title{make parameter for method of splitting factors (unordered categorical variables)}
\usage{
prune_par(values = c("aic", "bic"))
}
\arguments{
\item{trim_min}{An integer specifying the minimum value in the parameter grid}

\item{trim_max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make parameter for method of splitting factors (unordered categorical variables)
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.001, alpha_max = 0.1),
   prune_par(),
   size = 10
   )
}
