% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{ranefstart_par}
\alias{ranefstart_par}
\title{make parameter for determining swapping between tree and}
\usage{
ranefstart_par(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{a vector of possible values c("chisq", "max", "L2"), see ?mob_control()}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make parameter for determining swapping between tree and
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.001, alpha_max = 0.1),
   new_qual_param(),
   size = 10
   )
}
