% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{replace_prune_when_null}
\alias{replace_prune_when_null}
\title{internal function which takes the temporary tuning grid (used within cross validate)
and swaps the "null" as a character to it as NULL}
\usage{
replace_prune_when_null(tuning_grid_temp)
}
\description{
internal function which takes the temporary tuning grid (used within cross validate)
and swaps the "null" as a character to it as NULL
}
