% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{replace_ranefstart_when_false}
\alias{replace_ranefstart_when_false}
\title{internal function which takes the temporary tuning grid (used within cross validate)
and swaps the "null" as a character to it as NULL}
\usage{
replace_ranefstart_when_false(tuning_grid_temp)
}
\description{
internal function which takes the temporary tuning grid (used within cross validate)
and swaps the "null" as a character to it as NULL
}
