% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{restart_par}
\alias{restart_par}
\title{make parameter for determining restart parameter. TRUE = re-estimate downstream estimates of continuous
splitting variables when same variable is re-selected. FALSE = start at prior split value}
\usage{
restart_par(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{a vector of possible values; only TRUE/FALSE possible (the default)}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make parameter for determining restart parameter. TRUE = re-estimate downstream estimates of continuous
splitting variables when same variable is re-selected. FALSE = start at prior split value
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.10, alpha_max = 0.001),
   bonferroni_par(),
   size = 10
   )
}
