% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{rmse}
\alias{rmse}
\title{root mean square error of approximation}
\usage{
rmse(observed_y, predicted_y)
}
\arguments{
\item{observed_y}{A numeric vector of observed ys.}

\item{predicted_y}{A numeric vector of predictions of y.}
}
\value{
A double vector of length 1
}
\description{
root mean square error of approximation
}
\examples{
set.seed(1)
obs_y <- rnorm(4)
pred_y <- rnorm(4)
rmse(observed_y = obs_y, predicted_y = pred_y)
}
