% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_multilevel}
\alias{sim_multilevel}
\title{simulate tibble of two-level data with outcome,
id_vector, and categorical/continuous covariates with varied magnitudes of effect.
Nuisance predictors (orthogonal to outcome) also included.}
\usage{
sim_multilevel(
  j = 116,
  intercept_1 = 0,
  residual_var_sd_1 = 0.4,
  random_int_mean_2 = 0,
  residual_var_sd_2 = 0.1,
  start_seed = 713,
  num_in_groups = rpois(n = j, lambda = 15)
)
}
\arguments{
\item{j}{integer—number level 2 units}

\item{intercept_1}{double—intercept at level 1}

\item{residual_var_sd_1}{double—standard deviation of residuals at level 1}

\item{random_int_mean_2}{double—mean of random intercept at level 2}

\item{residual_var_sd_2}{double—standard deviation of random intercept at level 2,}

\item{start_seed}{integer—starting seed,}

\item{num_in_groups}{numeric vector—number of observations per group length of argument j}
}
\value{
tibble
}
\description{
simulate tibble of two-level data with outcome,
id_vector, and categorical/continuous covariates with varied magnitudes of effect.
Nuisance predictors (orthogonal to outcome) also included.
}
\examples{
   J <- 116
   dat <-
      sim_multilevel(
         j = J,
         intercept_1 = 0,
         residual_var_sd_1 = 0.7,
         random_int_mean_2 = 0,
         residual_var_sd_2 = 0.3,
         start_seed = 713,
         num_in_groups = rpois(n = J, lambda = 15)
        )
}
