% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_2021.R
\docType{data}
\name{spending_2021}
\alias{spending_2021}
\title{Data from Lisa Lendway's backyard garden}
\format{
A tibble with 18 rows and 5 variables:
\describe{
  \item{vegetable}{type of vegetable seed/plant OR other supply type, ie. dirt}
  \item{variety}{variety of vegetable seed/plant OR further description of other product}
  \item{brand}{brand of seed or plant}
  \item{price}{raw price - no taxes}
  \item{price_with_tax}{price with taxes}
}
}
\source{
Manually collected by Lisa Lendway, initially in a Google sheet.
}
\usage{
spending_2021
}
\description{
See documentation for the garden_spending dataset for more detail. This dataset summarizes how much was spent on gardening materials.
}
\keyword{datasets}
