% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{trim_par}
\alias{trim_par}
\title{make parameter for tuning level of amount of outlier trimmed in instability tests}
\usage{
trim_par(trim_min = 0.1, trim_max = 0.5)
}
\arguments{
\item{trim_min}{An integer specifying the minimum value in the parameter grid}

\item{trim_max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for maxdepth
}
\description{
make parameter for tuning level of amount of outlier trimmed in instability tests
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.001, alpha_max = 0.1),
   trim_par(trim_min = 0.1, trim_max = 0.3),
   size = 10
   )
}
