% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{alpha_par}
\alias{alpha_par}
\title{make parameter for tuning level of significance in instability tests}
\usage{
alpha_par(alpha_min = 0.05, alpha_max = 0.001)
}
\arguments{
\item{alpha_min}{An integer specifying the minimum value in the parameter grid}

\item{alpha_max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for maxdepth
}
\description{
make parameter for tuning level of significance in instability tests
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(min = 2, max = 5),
   alpha_par(min = 0.10, max = 0.001),
   trim_par(min = 0.1, max = 0.3),
   size = 10
   )
}
