% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\docType{data}
\name{catsplit_par}
\alias{catsplit_par}
\title{make parameter for method of splitting factors (unordered categorical variables)}
\format{
An object of class \code{qual_param} (inherits from \code{param}) of length 4.
}
\usage{
catsplit_par
}
\arguments{
\item{trim_min}{An integer specifying the minimum value in the parameter grid}

\item{trim_max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for catsplit
}
\description{
make parameter for method of splitting factors (unordered categorical variables)
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(maxdepth_min = 2, maxdepth_max = 5),
   alpha_par(alpha_min = 0.10, alpha_max = 0.001),
   catsplit_par(),
   size = 10
   )
}
\keyword{datasets}
