% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{correct_ranefstart}
\alias{correct_ranefstart}
\title{internal function which takes the temporary tuning grid (used within cross validate)
returns what glmertree expects}
\usage{
correct_ranefstart(tuning_grid_temp)
}
\description{
internal function which takes the temporary tuning grid (used within cross validate)
returns what glmertree expects
}
