% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_functions.R
\name{maxdepth_par}
\alias{maxdepth_par}
\title{make parameter for tuning maximum depth of tree}
\usage{
maxdepth_par(min = 2, max = 12)
}
\arguments{
\item{min}{An integer specifying the minimum value in the parameter grid}

\item{max}{An integer specifying the maximum value in the parameter grid}
}
\value{
A quantitative parameter, from \code{{dials}} for maxdepth
}
\description{
make parameter for tuning maximum depth of tree
}
\examples{
dials::grid_max_entropy(
   maxdepth_par(min = 2, max = 5),
   alpha_par(min = 0.10, max = 0.001),
   trim_par(min = 0.1, max = 0.3),
   size = 10
   )
}
